+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - added modelCallSettings property to onAgentStarted event for overriding AI SDK model call parameters
 - correctly handling multiple reasoning message parts
 - falling back to default agent profile when opening new project with agent profile that does not exist
+- clear exited aider process state
 
 ## [0.75.0]